refactor: move clap utils
