+++ bin/reth/src/args/mod.rs
@@ -24,4 +24,6 @@ mod stage_args;
 pub use stage_args::StageEnum;
 
 mod gas_price_oracle_args;
+pub mod utils;
+
 pub use gas_price_oracle_args::GasPriceOracleArgs;
