+++ bin/reth/src/debug_cmd/execution.rs
@@ -4,7 +4,7 @@ use crate::{
     dirs::{DataDirPath, MaybePlatformPath},
     node::events,
     runner::CliContext,
-    utils::{genesis_value_parser, get_single_header},
+    utils::get_single_header,
 };
 use clap::Parser;
 use futures::{stream::select as stream_select, StreamExt};
@@ -29,6 +29,7 @@ use reth_primitives::{stage::StageId, BlockHashOrNumber, BlockNumber, ChainSpec,
 use reth_provider::{providers::get_stage_checkpoint, ShareableDatabase, Transaction};
 use reth_staged_sync::utils::init::{init_db, init_genesis};
 
+use crate::args::utils::genesis_value_parser;
 use reth_stages::{
     sets::DefaultStages,
     stages::{
